+++ app/controllers/application_controller/ci_processing.rb
@@ -21,6 +21,8 @@ module ApplicationController::CiProcessing
     include Mixins::Actions::HostActions::Power
     include Mixins::Actions::HostActions::Misc
 
+    include Mixins::Actions::ProviderActions::MassTransform
+
     include Mixins::ExplorerShow
   end
 
@@ -1073,6 +1075,7 @@ module ApplicationController::CiProcessing
     when "instance_add_security_group"      then add_security_group_vms
     when "instance_remove_security_group"   then remove_security_group_vms
     when "vm_transform"                     then vm_transform
+    when "vm_transform_mass"                then vm_transform_mass
     end
   end
 end
