+++ spec/controllers/ems_infra_controller_spec.rb
@@ -84,10 +84,10 @@ describe EmsInfraController do
       expect(controller.send(:flash_errors?)).not_to be_truthy
     end
 
-    it "when vm_transform is pressed" do
+    it "when vm_transform_mass is pressed" do
       ems_infra = FactoryGirl.create(:ems_vmware)
-      expect(controller).to receive(:vm_transform)
-      post :button, :params => {:pressed => "vm_transform", :id => ems_infra.id, :format => :js}
+      expect(controller).to receive(:vm_transform_mass)
+      post :button, :params => {:pressed => "vm_transform_mass", :id => ems_infra.id, :format => :js}
       expect(controller.send(:flash_errors?)).not_to be_truthy
     end
   end
